lee 999

+++ MySuperApp/Class/Other/waterFllow/WaterFlowView.h
@@ -16,8 +16,8 @@
 
 @interface WaterFlowView : UIScrollView<UITableViewDelegate,UITableViewDataSource>{
 
-    int _columnCount;   //列数
-    int _cellsTotal;    //总的数据项
+    NSInteger _columnCount;   //列数
+    NSInteger _cellsTotal;    //总的数据项
     float _cellWidth;   //每列的宽度
 
     UIActivityIndicatorView *activityViewLoad;
@@ -28,8 +28,8 @@
     id<WaterFlowViewDataSource> waterFlowViewDatasource;
 }
 
-@property (nonatomic,assign) int columnCount; 
-@property (nonatomic,assign) int cellsTotal; 
+@property (nonatomic,assign) NSInteger columnCount;
+@property (nonatomic,assign) NSInteger cellsTotal;
 @property (nonatomic,assign) float cellWidth;
 @property (nonatomic,assign) id<WaterFlowViewDelegate> waterFlowViewDelegate;
 @property (nonatomic,assign) id<WaterFlowViewDataSource> waterFlowViewDatasource;
